+++ Mods ChoGGi/Example Layout Construction/metadata.lua
@@ -1,0 +1,16 @@
+return PlaceObj("ModDef", {
+	"title", "Example Layout Construction",
+	"version", 1,
+	"version_major", 0,
+	"version_minor", 1,
+	"saved", 0,
+	"image", "Preview.png",
+	"id", "ChoGGi_ExampleLayoutConstruction",
+	"author", "ChoGGi",
+	"lua_revision", 245618,
+	"code", {
+		"Code/Script.lua",
+	},
+	"TagBuildings", true,
+	"description", [[Add a layout building (Self-Sufficient Dome) to Infrastructure.]],
+})