+++ Mods ChoGGi/Fix Colonists Long Walks/Code/Script.lua
@@ -1,0 +1,43 @@
+-- See LICENSE for terms
+
+local dome_walk_dist = const.ColonistMaxDomeWalkDist
+
+-- local func copy pasta
+local function CheckDist(bld1, bld2)
+	local p1, p2 = ResolvePos(bld1, bld2), ResolvePos(bld2, bld1)
+	if not p1 or not p2 then
+		return false, max_int
+	end
+	if p1 == p2 then
+		return true, 0
+	end
+	local has_path
+	local len_sl = p1:Dist2D(p2)
+	if len_sl > dome_walk_dist then
+		return false, len_sl, true
+	end
+
+	local has_path, len = PathLenCached(p1, Colonist.pfclass, p2)
+
+	if has_path and len > dome_walk_dist then
+		has_path = false
+	end
+	return has_path or false, len
+end
+
+
+function IsInWalkingDistDome(bld1, bld2)
+	if bld1 == bld2 then
+		return true, 0
+	end
+	local dist_map = g_DomeToDomeDist[bld1]
+	local dist = dist_map and dist_map[bld2]
+	if dist then
+		-- check the actual dist instead of just saying sure go for it
+		local p = AreDomesConnectedWithPassage(bld1, bld2)
+			and bld1:GetDist2D(bld2) <= dome_walk_dist and true
+		return dist[1] or p, dist[2]
+--~ 		return dist[1] or AreDomesConnectedWithPassage(bld1, bld2), dist[2]
+	end
+	return CheckDist(bld1, bld2)
+end