+++ Mods ChoGGi/Notification Disable/metadata.lua
@@ -1,0 +1,19 @@
+return PlaceObj("ModDef", {
+	"title", "Notification Disable",
+	"version", 1,
+	"version_major", 0,
+	"version_minor", 1,
+	"saved", 0,
+	"id", "ChoGGi_NotificationDisable",
+	"author", "ChoGGi",
+	"steam_id", "1812220211",
+	"pops_any_uuid", "be6b458f-ad14-42dc-83cd-b4aa4b461fa2",
+	"code", {
+		"Code/Script.lua",
+	},
+	"image", "Preview.png",
+	"lua_revision", 245618,
+	"has_options", true,
+	"description", [[Use to disable certain notifications, look in Mod Options for the list.
+Red for Critical, and Blue for Important.]],
+})