+++ Mods ChoGGi/Passages Use Empty Hexes/metadata.lua
@@ -19,6 +19,7 @@ return PlaceObj("ModDef", {
 For now this is ONLY useful for the Medium, Mega, and Geoscape domes.
 I haven't figured out the grid connection funcs to be able to use the dome edges.
 Check the selection panel after placing one, if it says Connected to building then it'll work.
+See the screenshots above to know which hexes work.
 
 
 Known Issues: